Reworking context assembly to create additional indexsettings annotation.
I needed to break out the index settings from the query bean annotation
since indexing will be done with the resultbean object, and it would be
innappropriate to have a query bean annotation on a result object just
for index information.   This still isn't DRY enough, though, with
index settings potentially repeated in multiple places.